+++ script.ts/common/index.mts
@@ -1,2 +1,2 @@
 export * from './libs.mjs';
-export { Util } from './Util.mjs';
+export { Util } from './util.mjs';
